+++ config/disallowlister.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    /*
+     * Set an array of disallowed strings here. For example:
+     * 'disallowed_strings' => [
+     *     '*foo*',
+     *     'b?r',
+     *     'b[o,a]z'
+     * ],
+     */
+    'disallowed_strings' => [],
+];