+++ src/Facades/Disallowlister.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Accentinteractive\LaravelDisallowlister\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class Disallowlister extends Facade
+{
+
+    protected static function getFacadeAccessor()
+    {
+        return 'Disallowlister';
+    }
+}