+++ src/LaravelDisallowlister.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Accentinteractive\LaravelDisallowlister;
+
+class LaravelDisallowlister extends \Accentinteractive\DisallowLister
+{
+
+}