+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
 All notable changes to `laravel-disallowlister` will be documented in this file
+
+## 0.1.0 2021-08-23
+- Created accentinteractive/laravel-disallowlister
+- Added facade `Disallowlister`
+- Added publishable config file `disallowlister.php` to contain an default array of disallowed strings.